rc/OptionContainer.cpp: Fix segfault
Whe you provide 3 auth plugins (for example) and only 2 filter_ports e2guardian
crash with a SYSSEGV. We manage to fix this by testing if the number of filter_ports
is the same as the number of auth plugins.

This is a quick fix and maybe needs a deeper investigation.