Use automake as it was intended.
The only user visible change vis-a-vis e2guardian 3.4.0.1 is
that:

  share/e2guardian/languages/portuguese/messagesaccents
  share/e2guardian/languages/ukenglish/messages.alt

are intalled. If they shouldn't be, then they should be removed
from the tree.

  configs/lists/nocheckcertsitelist.in

has been removed as it was identical to nocheckcertsitelist.

  configs/authplugins/port.conf

has been removed and is properly generated from port.conf.in.

Aggressive use of DISTCLEANFILES is not useful, and was probably
a result of the handcrafted install-data rules (removed with
this commit) which prevented building in a separate objdir.