e2guardian -g causes segfault #163
This patch fixes the segfault caused by permission issues encountered
during a gentle reload of list files. It does this by adding a
condition to handle a stat return errno of EACCES by causing the
list to be reloaded. This should not result in any noticable
performance issues that I can see, although I have not benchmarked
it to see.