e2guardian -g causes segfault rev 2 #167
The cause of this version of the problem was the failure of the code
to unset the gentlereload flag when various issues are encountered
which require a full reload instead. This failure resulted in the
creation of an infinite loop in the main fc_controlit() loop.

This patch corrects this by resetting the gentlereload flag when
we determine a full reload is required.

This is probably the root of a number of segfault issues associated
with the -g switch.