Update template.html to a more modern look.
Updated template.html block page for a cleaner more modernised look, while adding in an image which is in base64 in order to reduce GET requests. User names are shown, as well as the category the blocked content is in. I initially created this for my home use, but due to a lot of requests for it on the pfSense forums, decided to submit it here.

Let me know if this is the correct branch or if you'd like me to submit to master.  Furthermore, if you have anything to change or add to this, feel free to. :)

As always, keep up the amazing work!