Add Filter Group and IP to template
Added filtergroup and IP to the template in order to make debugging much easier. And the information maybe useful in a lot of cases, and actually act as a warning to the user, that we 'know' who they are.

I've also changed around the order the blocked category, and reason is shown. I think it makes sense to see the category first.