Fix bug #841 - generated cert dates validaition produces 1970 end date with CA root notAfterDate greater than 2k38.
Turns out that it is not a time_t overflow issue within the e2g code.
The issue is with the openssl library ANS1_TIME_diff function which refuses to handle dates above 2k38.
It does returns an error, and so I have modified the code to check for this, so that the cacertenddate
is not changed if an error occurs in the ANS1_TIME_diff processes.